api/requirements.txt
requests==2.27.1
click==6.7
Flask==1.1.2
Flask-Migrate==3.1.0
authlib==1.0.1
Flask-Cors==3.0.10
Flask-HTTPAuth==4.0.0
Flask-SQLAlchemy==2.4.1
itsdangerous==0.24
Jinja2==2.11.2
MarkupSafe==1.1.1
PyJWT==1.7.1
SQLAlchemy==1.3.6
Werkzeug==1.0.1
